+++ pmd-matlab/src/test/java/net/sourceforge/pmd/cpd/MatlabTokenizerTest.java
@@ -54,4 +54,36 @@ public class MatlabTokenizerTest extends AbstractTokenizerTest {
         TokenEntry.getEOF();
         assertEquals(2, tokens.size()); // 2 tokens: "end" + EOF
     }
+
+    @Test
+    public void testComments() throws IOException {
+        SourceCode sourceCode = new SourceCode(new SourceCode.StringCodeLoader("classdef LC" + PMD.EOL
+                + "    methods" + PMD.EOL
+                + "        function [obj, c,t, s ] = Classification( obj,m,t, cm )%#codegen" + PMD.EOL
+                + "        end" + PMD.EOL
+                + "    end" + PMD.EOL
+                + "end"));
+        Tokens tokens = new Tokens();
+        tokenizer.tokenize(sourceCode, tokens); // should not result in parse error
+        TokenEntry.getEOF();
+        assertEquals(28, tokens.size());
+    }
+
+    @Test
+    public void testBlockComments() throws IOException {
+        SourceCode sourceCode = new SourceCode(new SourceCode.StringCodeLoader("%{" + PMD.EOL
+                + "  Name:     helloworld.m\n" + PMD.EOL
+                + "  Purpose:  Say \"Hello World!\" in two different ways" + PMD.EOL
+                + "%}" + PMD.EOL
+                + PMD.EOL
+                + "% Do it the good ol' fashioned way...command window" + PMD.EOL
+                + "disp('Hello World!');\n" + PMD.EOL
+                + "%" + PMD.EOL
+                + "% Do it the new hip GUI way...with a message box" + PMD.EOL
+                + "msgbox('Hello World!','Hello World!');"));
+        Tokens tokens = new Tokens();
+        tokenizer.tokenize(sourceCode, tokens); // should not result in parse error
+        TokenEntry.getEOF();
+        assertEquals(13, tokens.size());
+    }
 }